Removed Additional C++ Programs Folder and merged it's contents into Hacktoberfest-2021/C++-Programs
